step 2. 문자열 계산기 리뷰 반영. expression 분리
