+++ src/main/kotlin/CalculatorSimulator.kt
@@ -4,6 +4,6 @@ fun main() {
     print("please input calculation expression :")
     val expression = readln()
     val calculator = Calculator(expression)
-    val result = calculator.calculate()
+    val result = calculator.result()
     println(expression + " = " + result)
 }
